refactor(#8): 이미지 메타데이터를 찾지 못해 삭제하지 못했을 때 404 에러를 반환하도록 함

+++ src/main/java/com/aetheri/application/service/image/DeleteImageMetadataService.java
@@ -2,6 +2,8 @@ package com.aetheri.application.service.image;
 
 import com.aetheri.application.port.in.image.DeleteImageMetadataUseCase;
 import com.aetheri.domain.adapter.out.r2dbc.ImageMetadataRepositoryR2dbcAdapter;
+import com.aetheri.domain.exception.BusinessException;
+import com.aetheri.domain.exception.message.ErrorMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -15,7 +17,17 @@ public class DeleteImageMetadataService implements DeleteImageMetadataUseCase {
 
     @Override
     public Mono<Void> deleteImageMetadata(Long runnerId, Long imageId) {
-        return imageMetadataRepositoryR2DbcAdapter.deleteById(runnerId, imageId)
+        return imageMetadataRepositoryR2DbcAdapter.isExistImageMetadata(imageId)
+                .flatMap(isExist -> {
+                    if (isExist) return imageMetadataRepositoryR2DbcAdapter.deleteById(runnerId, imageId);
+
+                    return Mono.error(
+                            new BusinessException(
+                                    ErrorMessage.NOT_FOUND_IMAGE_METADATA,
+                                    "요청한 이미지 메타데이터를 찾지 못해 삭제하지 못했습니다."
+                            )
+                    );
+                })
                 .doOnSuccess(l -> log.info("[DeleteImageService] 사용자 {}에 의해 이미지 {}개가 삭제되었습니다.", runnerId, l))
                 .then();
     }
